CPT-80 locale: add temp i18n solution for @pingcap-community/ui
