+++ packages/ui/src/components/footer/footer.data.js
@@ -1,77 +1,9 @@
-export const navItems = [
-  {
-    title: 'Community',
-    items: [
-      {
-        title: 'TiKV',
-        link: 'https://tikv.org',
-      },
-      {
-        title: 'Chaos Mesh®',
-        link: 'https://chaos-mesh.org',
-      },
-      {
-        title: 'Slack',
-        link: 'https://slack.tidb.io/invite?team=tidb-community&channel=everyone',
-      },
-      {
-        title: 'Forum',
-        link: 'https://asktug.com',
-      },
-      {
-        title: 'Mailing List',
-        link: 'https://lists.tidb.io/g/main/subgroups',
-      },
-    ],
-  },
-  {
-    title: 'Resources',
-    items: [
-      {
-        title: 'Quick Start',
-        link: 'https://docs.pingcap.com/zh/tidb/stable/quick-start-with-tidb',
-      },
-      {
-        title: 'Best Practices',
-        link: 'https://docs.pingcap.com/zh/tidb/stable/tidb-best-practices',
-      },
-      {
-        title: 'FAQs',
-        link: 'https://docs.pingcap.com/zh/tidb/stable/tidb-faq',
-      },
-      {
-        title: 'Release Notes',
-        link: 'https://docs.pingcap.com/zh/tidb/dev/release-notes',
-      },
-      {
-        title: 'Blog',
-        link: 'https://pingcap.com/blog',
-      },
-      {
-        title: 'Github',
-        link: 'https://github.com/pingcap/tidb',
-      },
-    ],
-  },
-  {
-    title: 'Learning',
-    items: [
-      {
-        title: 'Docs',
-        link: 'https://docs.pingcap.com/tidb/stable',
-      },
-      {
-        title: 'Case Studies',
-        link: 'https://pingcap.com/case-studies',
-      },
-      {
-        title: 'PingCAP University',
-        link: 'https://university.pingcap.com',
-      },
-      {
-        title: 'TiDB in Action',
-        link: 'https://book.tidb.io',
-      },
-    ],
-  },
-];
+import * as dataEn from './footer.data-en';
+import * as dataCn from './footer.data-Cn';
+
+export default {
+  en: dataEn,
+  cn: dataCn,
+};
+
+export * from './footer.data-en';
