+++ packages/ui/src/components/header/header.data-cn.js
@@ -1,0 +1,133 @@
+export const navItems = [
+  {
+    title: '关于',
+    items: [
+      {
+        title: '社区行为准则',
+        link: 'https://github.com/pingcap/community/blob/master/CODE_OF_CONDUCT.md',
+      },
+      {
+        title: 'TiDB 和 PingCAP',
+        link: 'https://pingcap.com',
+      },
+    ],
+  },
+  {
+    title: '用户组',
+    items: [
+      {
+        title: '用户组首页',
+        link: 'https://tug.tidb.io',
+      },
+      {
+        title: 'MVA',
+        link: 'https://tug.tidb.io/MVA',
+      },
+    ],
+  },
+  {
+    title: '开发者社区',
+    items: [
+      {
+        title: '开发者社区首页',
+        link: 'https://contributor.tidb.io',
+      },
+      {
+        title: '如何参与贡献',
+        items: [
+          {
+            title: '源码阅读系列文章',
+            link: 'https://pingcap.com/blog-cn/#TiDB-%e6%ba%90%e7%a0%81%e9%98%85%e8%af%bb',
+          },
+          {
+            title: 'Talent Plan',
+            link: 'https://university.pingcap.com/talent-plan/',
+          },
+        ],
+      },
+      {
+        title: '人员',
+        items: [
+          {
+            title: 'Maintainers',
+            link: 'https://contributor.tidb.io/people/maintainer',
+          },
+          {
+            title: 'Committer',
+            link: 'https://contributor.tidb.io/people/committer',
+          },
+          {
+            title: 'Reviewer',
+            link: 'https://contributor.tidb.io/people/reviewer',
+          },
+          {
+            title: 'Active Contributor',
+            link: 'https://contributor.tidb.io/people/active-contributor',
+          },
+          {
+            title: 'Contributor',
+            link: 'https://contributor.tidb.io/people/contributor',
+          },
+        ],
+      },
+      {
+        title: '特殊兴趣小组(SIG)',
+        link: 'https://contributor.tidb.io/SIG/',
+      },
+    ],
+  },
+  {
+    title: '学习',
+    items: [
+      {
+        title: 'TiDB DBA 课程',
+        link: 'https://university.pingcap.com/',
+      },
+      {
+        title: 'Talent Plan 课程',
+        link: 'https://university.pingcap.com/talent-plan/',
+      },
+      {
+        title: 'PingCAP Certified TiDB Associate (PCTA)',
+        link: 'https://university.pingcap.com/certificate/PCTA/',
+      },
+      {
+        title: 'PingCAP Certified TiDB Professional (PCTP)',
+        link: 'https://university.pingcap.com/certificate/PCTP/',
+      },
+    ],
+  },
+  {
+    title: '活动',
+    link: 'https://contributor.tidb.io/events',
+  },
+  {
+    title: '资源',
+    items: [
+      {
+        title: '论坛',
+        link: 'https://asktug.com',
+      },
+      {
+        title: 'Slack',
+        link: 'https://slack.tidb.io/invite?team=tidb-community&channel=everyone',
+      },
+      {
+        title: '文档',
+        link: 'https://docs.pingcap.com/tidb/stable',
+      },
+      {
+        title: '博客',
+        link: 'https://asktug.com/c/blog/l/latest',
+      },
+      {
+        title: '邮件列表',
+        link: 'https://lists.tidb.io/g/main/subgroups',
+      },
+    ],
+  },
+  {
+    title: '找工作',
+    link: 'https://pingcap.com/community-cn/careers/join/',
+  },
+];
